fix: fix ignore file on windows when used with linux then
