+++ packages/cli/src/__tests__/utils.test.ts
@@ -1,25 +1,6 @@
-import { isSubdir, slash } from '../utils';
+import { isSubdir } from '../utils';
 
 jest.mock("os");
-
-
-describe('slash path', () => {
-  it('can correctly slash path', () => {
-    [
-      ['foo\\bar', 'foo/bar'],
-      ['foo/bar', 'foo/bar'],
-      ['foo\\中文', 'foo/中文'],
-      ['foo/中文', 'foo/中文'],
-    ].forEach(([path, expectRes]) => {
-      expect(slash(path)).toBe(expectRes);
-    });
-  });
-
-  it('does not modify extended length paths', () => {
-    const extended = '\\\\?\\some\\path';
-    expect(slash(extended)).toBe(extended);
-  });
-});
 
 describe('isSubdir', () => {
   it('can correctly determine if subdir', () => {
