+++ packages/cli/src/commands/push.ts
@@ -5,7 +5,7 @@ import { performance } from 'perf_hooks';
 import { yellow, green, blue } from 'colorette';
 import { createHash } from 'crypto';
 
-import { bundle, Config, loadConfig, RedoclyClient, IGNORE_FILE, BundleOutputFormat, getTotals } from '@redocly/openapi-core';
+import { bundle, Config, loadConfig, RedoclyClient, IGNORE_FILE, BundleOutputFormat, getTotals, slash } from '@redocly/openapi-core';
 import {
   promptUser,
   exitWithError,
@@ -13,7 +13,6 @@ import {
   getFallbackEntryPointsOrExit,
   pluralize,
   dumpBundle,
-  slash
 } from '../utils';
 
 type Source = {
