+++ packages/cli/src/utils.ts
@@ -286,19 +286,6 @@ export function exitWithError(message: string) {
   process.exit(1);
 }
 
-
-/**
- * Convert Windows backslash paths to slash paths: foo\\bar ➔ foo/bar
- */
-export function slash(path: string): string {
-  const isExtendedLengthPath = /^\\\\\?\\/.test(path)
-  if (isExtendedLengthPath) {
-    return path
-  }
-
-  return path.replace(/\\/g, '/');
-}
-
 /**
  * Checks if dir is subdir of parent
  */
