+++ packages/core/src/__tests__/utils.test.ts
@@ -1,4 +1,4 @@
-import { pickObjectProps, omitObjectProps } from '../utils';
+import { pickObjectProps, omitObjectProps, slash } from '../utils';
 
 describe('utils', () => {
   const testObject = {
@@ -51,6 +51,24 @@ describe('utils', () => {
 
     it('returns an empty object if empty target obj passed', () => {
       expect(omitObjectProps({}, ['d', 'e'])).toStrictEqual({});
+    });
+  });
+
+  describe('slash path', () => {
+    it('can correctly slash path', () => {
+      [
+        ['foo\\bar', 'foo/bar'],
+        ['foo/bar', 'foo/bar'],
+        ['foo\\中文', 'foo/中文'],
+        ['foo/中文', 'foo/中文'],
+      ].forEach(([path, expectRes]) => {
+        expect(slash(path)).toBe(expectRes);
+      });
+    });
+
+    it('does not modify extended length paths', () => {
+      const extended = '\\\\?\\some\\path';
+      expect(slash(extended)).toBe(extended);
     });
   });
 });
