+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export { BundleOutputFormat, readFileFromUrl } from './utils';
+export { BundleOutputFormat, readFileFromUrl, slash } from './utils';
 export { Oas3_1Types } from './types/oas3_1';
 export { Oas3Types } from './types/oas3';
 export { Oas2Types } from './types/oas2';
