+++ packages/core/src/rules/utils.ts
